+++ packages/sdk/electron/src/ElectronDataManager.ts
@@ -14,7 +14,6 @@ import {
   readFlagsFromBootstrap,
 } from '@launchdarkly/js-client-sdk-common';
 
-import type { ElectronIdentifyOptions } from './ElectronIdentifyOptions';
 import type { ValidatedOptions } from './options';
 
 const logTag = '[ElectronDataManager]';
@@ -68,12 +67,11 @@ export default class ElectronDataManager extends BaseDataManager {
 
     // When bootstrap is provided, we will resolve the identify immediately. Then we will fallthrough to connect
     // to the configured connection mode.
-    const electronIdentifyOptions = identifyOptions as ElectronIdentifyOptions | undefined;
-    if (electronIdentifyOptions?.bootstrap) {
-      this._finishIdentifyFromBootstrap(context, electronIdentifyOptions, identifyResolve);
+    if (identifyOptions?.bootstrap) {
+      this._finishIdentifyFromBootstrap(context, identifyOptions, identifyResolve);
     }
     // Bootstrap path already called resolve so we use this to prevent duplicate resolve calls.
-    const resolvedFromBootstrap = !!electronIdentifyOptions?.bootstrap;
+    const resolvedFromBootstrap = !!identifyOptions?.bootstrap;
 
     const offline = this.connectionMode === 'offline';
     // In offline mode we do not support waiting for results.
@@ -109,12 +107,12 @@ export default class ElectronDataManager extends BaseDataManager {
 
   private _finishIdentifyFromBootstrap(
     context: Context,
-    electronIdentifyOptions: ElectronIdentifyOptions,
+    identifyOpts: LDIdentifyOptions,
     identifyResolve: () => void,
   ): void {
-    let { bootstrapParsed } = electronIdentifyOptions;
+    let { bootstrapParsed } = identifyOpts;
     if (!bootstrapParsed) {
-      bootstrapParsed = readFlagsFromBootstrap(this.logger, electronIdentifyOptions.bootstrap);
+      bootstrapParsed = readFlagsFromBootstrap(this.logger, identifyOpts.bootstrap);
     }
     this.flagManager.setBootstrap(context, bootstrapParsed);
     this._debugLog('Identify - Initialization completed from bootstrap');
@@ -129,7 +127,7 @@ export default class ElectronDataManager extends BaseDataManager {
   ) {
     const rawContext = Context.toLDContext(context)!;
 
-    const plainContextString = JSON.stringify(Context.toLDContext(context));
+    const plainContextString = JSON.stringify(rawContext);
     const requestor = makeRequestor(
       plainContextString,
       this.config.serviceEndpoints,